+++ Runtime/async/timer.cs
@@ -15,10 +15,10 @@ namespace Unity.UIWidgets.async {
             this.cancel();
         }
 
-        public static double timeSinceStartup {
+        public static float timeSinceStartup {
             get {
 #if UNITY_EDITOR
-                return EditorApplication.timeSinceStartup;
+                return (float) EditorApplication.timeSinceStartup;
 #else
                 return Time.realtimeSinceStartup;
 #endif
@@ -143,7 +143,7 @@ namespace Unity.UIWidgets.async {
         }
 
         class TimerImpl : Timer, IComparable<TimerImpl> {
-            double _deadline;
+            float _deadline;
             readonly Action _callback;
             bool _done;
 
@@ -151,7 +151,7 @@ namespace Unity.UIWidgets.async {
             readonly TimeSpan _interval;
 
             public TimerImpl(TimeSpan duration, Action callback, bool periodic = false) {
-                this._deadline = timeSinceStartup + duration.TotalSeconds;
+                this._deadline = timeSinceStartup + (float) duration.TotalSeconds;
                 this._callback = callback;
                 this._done = false;
 
@@ -167,7 +167,7 @@ namespace Unity.UIWidgets.async {
                 this._done = false;
             }
 
-            public double deadline {
+            public float deadline {
                 get { return this._deadline; }
             }
 
@@ -197,7 +197,7 @@ namespace Unity.UIWidgets.async {
                 }
 
                 if (this.periodic) {
-                    this._deadline = now + this._interval.TotalSeconds;
+                    this._deadline = now + (float) this._interval.TotalSeconds;
                 }
             }
